Added a better reset fields. Ressurected show on monitor.
