+++ BluescreenSimulator/Views/MainWindow.xaml
@@ -27,7 +27,7 @@
     </Window.Resources>
     <Grid>
 
-        <TabControl Grid.Row="0" Name="Tabs" ItemsSource="{Binding Bluescreens}" SelectedItem="{Binding SelectedBluescreen}" SelectedIndex="1">
+        <TabControl Grid.Row="0" Name="Tabs" ItemsSource="{Binding Bluescreens}" SelectedItem="{Binding SelectedBluescreen}">
             <TabControl.ItemTemplate>
                 <DataTemplate DataType="{x:Type viewModels:IBluescreenViewModel}">
                     <TextBlock Text="{Binding StyleName}"/>
@@ -230,7 +230,7 @@
             <TextBox Grid.Row="3" Grid.Column="1" Name="CmdCommand" Text="{Binding SelectedBluescreen.CmdCommand}"/>
         </Grid>
         <UniformGrid Grid.Row="2" Margin="0 0" VerticalAlignment="Center" Rows="1" Columns="4">
-            <Button Command="{Binding SelectedBluescreen.ResetAllCommand}" Margin="10 0">Reset fields</Button>
+            <Button Command="{Binding ResetAllCommand}"  Margin="10 0">Reset fields</Button>
             <Button Click="OpenAbout" Margin="10 0">Instructions</Button>
             <Button Click="GenerateExe" Margin="10 0">Generate EXE/Cmd</Button>
             <Button Click="ShowBSOD" Margin="10 0" Visibility="{Binding SelectedBluescreen.IsNotWaiting, Converter={StaticResource VisibilityConverter}}">Give me an error</Button>
